DispatchQueue: only dequeue entries between headPtr and dispatchPtr
Case: vvvdvvvddddd
            ^    ^
If the leftmost instruction is a store and it writebacks, these instructions won't be replayed.
However, we cannot move headPtr to the left hand side of the dispatchPtr since there're still
instructions that have not been dispatched to issue queues.
In this case, we only remove the instructions before dispatchPtr.
Moving headPtr in case of store writeback only affects performance, since instructions leave dispatch queue when they commit.