Makefile: enable SRAM randomization for verilator simulation
Previously we don't use the --infer-rw and --repl-seq-mem flags for simulation verilog.
However, the SyncReadMem fails to generate random read data when ren is not set.
In this commit, SyncReadMem is changed to blackboxes and generated by the vlsi_mem_gen script.
RANDOMIZE_GARBAGE_ASSIGN flag is defined to enable randomization.