Makefile: enable SRAM randomization for verilator simulation (#739)
* ICache: set holdRead to true for meta and data SRAMs

SyncReadMem generates the verilog behavior model whose output rdata is always
mem(RegNext(raddr)). Accidentally, ICache will not change meta and data
SRAMs' raddr if the second pipeline stage is stalled (and ren is false).
Thus, the SRAMs seem to have the holdRead property.
Obviously, it will cause errors on real SRAMs. We set holdRead to true to fix the bug.

* L1plusCache: set holdRead to true for SRAMs

* Makefile: enable SRAM randomization for verilator simulation

Previously we don't use the --infer-rw and --repl-seq-mem flags for simulation verilog.
However, the SyncReadMem fails to generate random read data when ren is not set.
In this commit, SyncReadMem is changed to blackboxes and generated by the vlsi_mem_gen script.
RANDOMIZE_GARBAGE_ASSIGN flag is defined to enable randomization.