LoadUnit: block fast wakeup and writeback when killed in s1 (#777)
MMIO should never hit in DCache. However, DCache does not guarantee the miss
is strictly according to the vaddr, paddr provided by the pipeline, when the
load is killed. That is, DCache may response valid = false and miss = false
when the MMIO instruction is accessed and then killed.

For instructions with exceptions, it will cause flush when it retires and
not cause errors.