backend: add more instruction fusion cases
This commit adds more instruction fusion cases that need changes in both
the decode stage and the funtion units (ALU only now).

Previously we only add the fusion cases that can be fused into a RV64GCB
instruction. In this commit, we add some opcode to the function units
and fuse the new instruction pairs into these new internal uops.

The list of opcodes we add in this commit is shown below:

* szewl1: `slli r1, r0, 32` + `srli r1, r0, 31`

* szewl2: `slli r1, r0, 32` + `srli r1, r0, 30`

* byte2: `srli r1, r0, 8` + `andi r1, r1, 255`

* sh4add: `slli r1, r0, 4` + `add r1, r1, r2`

* sr30add: `srli r1, r0, 30` + `add r1, r1, r2`

* sr31add: `srli r1, r0, 31` + `add r1, r1, r2`

* sr32add: `srli r1, r0, 32` + `add r1, r1, r2`

* oddadd: `andi r1, r0, 1`` + `add r1, r1, r2`

* oddaddw: `andi r1, r0, 1`` + `addw r1, r1, r2`