arbiter: better balance among function units
This commit changes the splitN algorithm for the write-back arbiter.

Previously we split the function units as follows:
(FU0 FU1 FU2) (FU3 FU4 FU5).
However, this strategy tends to group the function units with the same
type into the same arbiter and may cause performance loss.

In this commit, we change the strategy to: (FU0 FU2 FU4) (FU1 FU3 FU5).