csr: add vectored trap mode (#1343)
All bits for stvec and mtvec are writable in XiangShan.

According to the RISC-V spec, {m,s}tvec[1:0] are MODE bits. When
MODE=Vectored, all synchronous exceptions into M/S mode cause the pc
to be set to the address in the BASE field, whereas interrupts cause
the pc to be set to the address in the BASE field plus four times
the interrupt cause number.

If XiangShan decides to not support vectored mode, {m,s}tvec[1:0]
should be hardwired to zero.