csr: update mtval/stval according to the trap mode
This commit changes the condition to update mtval and stval.

According to the RISC-V spec, when a trap is taken into M/S-mode,
mtval/stval is either set to zero or written wrih exception-specific
information to assist software in handling the trap.

Previously in XiangShan, mtval/stval is updated depending on the
current priviledge mode, which is incorrect.