rs: don't update midResult when flushed (#1758)
This commit fixes a bug when FMA partially issues but is flushed
just after it is issues. In this case, new instruction will enter
the RS and writes the data array. However, previously midResult
from FMA is written into the data array two cycles after issue.
This may cause the wrong data to be written into the data array.

This is a rare case because usually instructions enter RS in-order,
unless dispatch2 is blocked.