csr: fix interrupt number when updating cause
Rob detects interrupts before CSRs are updated. This does not cause
errors in single core, because interrupts will not change in these
cycles. However, in multi-core, interrupts may be cleared by other
cores, resulting in updating the mcause with zero interrupts. This
would cause errors in the software as the interrupt zero is reserved.