exu: fix the definition/instance in multicore (#2338)
Currently the definitions are instantiated every time when FUBlock
is called. Though it addresses the dedup issue for multiple ExeUnits
in one FUBlock, it still causes non-dedup modules across various
FUBlocks and multiple CPU cores.

* exu: do not use definition/instance for CSR

Difftest relies on the number of instantiations to determine the
numCores and number of bundles. Definition/Instance breaks this
requirement. We need to disable using definition/instance for CSR.
Since CSR will not be re-used via Instance, we keep it the same
as before for now.