LSU, Loadunit: Set all data to 0 when exception (#2888) (#2889)
When an exception occurs, as s3_out is true, backend will still consider
the data valid at this time (which is actually in X state) valid and
bypass it. Although in real chips, this situation will be handled after
2 cycles of exception processing (flushing the pipeline and
redirecting), when using vcs simulation, the simulation cannot continue
due to the problem of X-state propagation. Therefore, when an exception
occurs in load_s3, the output data will be forcibly set to 0.

Co-authored-by: Haoyuan Feng <fenghaoyuan19@mails.ucas.ac.cn>