Bump ready-to-run to update linux
The linux-hello test in CI has switched to openSBI in #2836. However, we
did not update ready-to-run at that time for some reason.

ready-to-run should be abandoned. Using git repo to manage prebuilt
binary files is not a good pratice and would cause repository bloat.
Also, lack of building infomation of binary files makes difficulties in
rebuilding them.