difftest: Support generating difftest bundles for `make verilog` target
Previously, difftest bundles can only be generated when building `sim-verilog`
target. The difftest can not be used when simulating design that use the output
of `make verilog` target.

This patch enable generating of difftest bundles for `make verilog` target,
it can further be used to enable difftest support in simulation.

Signed-off-by: Jiuyue Ma <majiuyue@bosc.ac.cn>