ICache: store needflush if a miss request is about to fire
A quick fix for CI https://github.com/OpenXiangShan/XiangShan/actions/runs/9221230229/job/25377821362\?pr\=3004 failure due to a cache multi-hit.

This happens when a fence.i came just the same cycle with a mainPipe toMSHR.fire.

In this case, MSHR will not store needflush since `state === s_idle`, and send miss request to L2 cache. When the response from L2 came, MSHR will refill data to data/meta SRAM, which should not happen.

Maybe a better way to do this is canceling request to L2 when a fence.i comes. However, this requires relatively large modifications to state machines in both mainPipe and MSHR, so is not adopted now.