difftest: check load only when isAmo or isLoad and skip MMIO
Previous we submit loadEvent for all committed instructions.
According to cpp code, we only check load when isAmo and isLoad.
And MMIO will be skipped, no need to check REF and GoldenMem.

When some loadEvent are dropped to speed up difftest, this change
fix legacy impact of loadEvent unprocessed before.