StoreQueue: Commit storequeue entry by judging robidx (#3020)
In the previous design, for scalar store commands, commits were made
based on the `scommit` passed in by rob. That is, after rob commits n
store commands in one cycle, it sets scommit to n, and the storequeue
needs to commit n entries based on the value of scommit.

Due to the special nature of vector instructions, a vector store
instruction may correspond to a maximum of 128 items, and it is not
possible to use `scommit` commits, which are similar to scalar commits.

In this modification, both vector and scalar directives are compared
against `robidx`, and the storequeue entry is committed when the `uop`
stored in `uop` is equal to or older than the `pendingptr` passed in by
rob. This has some improvement on timing.

Also, additional processing is currently required for vector
instructions. When an exception occurs on an item, it needs to be
handled with an out-of-queue, but not write sbuffer.