MISC: turn to use issue form to avoid improper filling
Since the last time we update our Issue Template, we found that there is a lot of people who did not fill it properly and even remove the whole template. To avoid that, this patch turns to use the beta issue form. We hope this will improve our issue quality.