StoreQueue: Add exception handling logic for vector instructions.
All vector instruction uop normally dequeue, but the Uop after the exception is raised does not write to the 'sbuffer'.
Flags are used to record whether there are any exceptions when the queue is displayed.
This is determined each time a write is made to the 'databuffer', prevent subsequent uop of the same instruction from writing to the 'dataBuffer'.