Bump CoupledL2 (#3148)
* MSHR: fix bug in DBID field of CompData with DCT

In a snoop DCT transaction, the RN-F provides the read data in CompData
that is sent directly to peer RN-F and the DBID must be set to the same
value as the TxnID of the snoop.

* MSHR: fix bug when a retried release writes DS repeatedly

Before a (TL) Release or (CHI) WriteBackFull/Evict request is sent to
TXREQ channel, the request is reponsible for writing the data to be
refilled into data storage. From that time on the incoming refill
request can continue to choose the same victim way. However in the CHI
version, WriteBackFull/Evict might retry after a while and when retry
happens we should avoid the task refills data into data storage
repeatedly.

Therefore in this commit, WriteBackFull/Evict is sent out towards main
pipe the first time with `AllowRetry = 1` in order to refill DS. In the
second time it will be sent to TXREQ channel directly with `AllowRetry =
0` to avoid repeated refill.