fix(L2TLB): Fix exception generation logic
We may currently generate three types of exceptions, pf, gpf, and af. There must be only one type of exception that should occur in each resp returned by L2 TLB, which is the type of exception that occurs for the first time during the PTW process. Among them
pf & gpf: the two cases correspond to stage1 and stage2 respectively. **In our previous design, the error is that onlyStage1 is also considered to need gpf checking, but in fact, onlyStage1 shouldn't report gpf**
af: there are two kinds of access faults, the first one is the access fault obtained by querying pmp before PTW accesses the memory, and the second one is the access fault obtained by the PPN high level of page table is not 0 after PTW accesses the memory. we call these two kinds of access faults as pmp_af and ppn_af respectively.

For allStage case: pf, gpf, af can happen. pf precedes gpf (if pf is reported in the first stage, it should be returned directly without checking gpf in the second stage). For af, if it's pmp_af, this af will be reported before actually accessing memory, and will have a higher priority than pf or gpf (actually, if pmp_af occurs, no memory will be accessed, and there will not be a pf or gpf at the same time). In case of ppn_af, this af should actually be checked in pmp before being reported before using this physical address for fetch or access. However, since our physical address will be truncated directly on return, we need to check the af in advance, and this af will have the lowest priority and will be lower than pf | gpf. (i.e., pf and gpf will not occur at the same time, pf > gpf. The two kinds of pf and pmp_af will not occur at the same time, but may occur at the same time as ppn_af, pmp_af > {pf or gpf} > ppn_af).

For onlyStage1: only pf or af will appear, same as above.
For onlyStage2: only gpf or af will appear, same as above.
For noS2xlate: only pf or af will appear, same as above.