fix(LoadUnit): forward resp is only valid when req valid (#3475)
In load_s1 stage, a forward request is sent to storequeue to check
whether there is any data forward, and also whether a violation occurs
when mdp turns on.

In storequeue forward check, both vaddr and paddr need to be checked, so
it needs TLB hit (!miss), and when TLB misses, the forward.valid is
false. When forward.valid is false, the forward checking information
returned by storequeue in load_s2 stage (for example, addrInvalid,
dataInvalid, etc.) is uncertain. Only when forward.valid signal of
load_s1 is true, can we judge the load replay condition based on the
forward information returned from storequeue in load_s2.

Therefore, we add the RegNext(io.lsq.forward.valid) condition to the
generation of s2_mem_amb and s2_fwd_fail signals, which are only
meaningful when RegNext(io.lsq.forward.valid) is true.