fix(dcache): fix perf bug of BankedDataArray
If the addresses(for example:0x80000000, 0x90000000) of two read requests fall in the same dcache set(0), the same bank(0), and different ways,
bank conflict will occur in the previous design. In fact, in the design of BankedDataArray, each read request will read all the way
of an entire bank. So this situation should not necessarily produce a bank conflict.