submodule(CoupledL2): fix bugs in PCredit management
This commit fixes the following possible PCredit related bugs:

1. RetryAck is typically sent by the Completer before PCrdGrant.
However, it is permitted to send RetryAck after PCrdGrant. When the
latter situation happens, the transaction with credit must only be sent
by the Requester after both the RetryAck response and an appropriate
PCrdGrant response are received. Before this fix CoupledL2 is unable to
handle the latter situation and might drop a PCredit or handle out a
PCredit to more than one MSHR.

2. The TxnID field of the PCrdGrant response is not used and must be
set to zero. However, before this fix, CoupledL2 hands out all the
RXRSP responses, including PCrdGrant, according to the most significant
bit in TxnID, which makes MMIOBridge unable to receive PCrdGrant.

This pull request fixes the above bugs by managing all the PCredit
among MMIO and cacheable MSHRs together. All the PCrdGrant responses
that CoupledL2 receives will be cached in `PCrdTypes` and `PCrdSrcIDs`
register files. When an MSHR receives a RetryAck, it will look for a
PCrdGrant with appropriate PCrdType and SrcID.