fix(LoadUnit): prefetch.i should not use HWPrefetch paddr
In previous design, prefetch.i would use HWPrefetch's paddr in load_s0 and send it to pmp in load_s1. However, when prefetch.i is valid, HWPrefetch's valid is 0 and paddr is X-stated, resulting in X-state propagation.

In fact, prefetch.i just sends vaddr to Frontend and does not use the result of the pmp check. In this commit, paddr that prefetch.i sends to pmp is written directly to 0, thus avoiding X-state propagation.