fix(PageTableCache): fix ptwcache refill logic when exception
Since L2 Cache access width is 512 bits, 8 PTE items are refilled into page cache together. In previous design, when an exception occurs in one of the 8 refilled PTE entries, none of the 8 entries will be refilled. In this case, if there is a normal entry in one of the 8 neighbouring page tables, it will not be refillled because of the exception in the other entries, which will lead to blocking.

This commit fixes this problem. `vs` means that if one of the 8 consecutive page tables has an exception, the corresponding `vs` will be forced to be false when judging hit. `onlypf` means that if x2slate is not enabled, and only pf occurs, then the pf information will be returned. Except for the onlypf case, the rest of the exceptions will not be refilled in page cache (none of them will hit).