fix(StoreQueue): fix load - vectorstore forward logic
In previous design, `addrValidVec` and `dataValidVec` in storequeue are only considered scala store cases, which means that `ValidVec` of a vector store will always be false, resulting in `forward.valid` always false. When the vector store commit, `allocted` will be set to false, loadqueuereplay cannot wait for the `Valid` signal, and be stuck.

[PR 3590](https://github.com/OpenXiangShan/XiangShan/pull/3590) fixes it.  In this commit, `ValidVec` in storequeue of vectore stores will be true when `vecMbCommit`, in order to imporve the performance of load - vectorstore forward.