fix(CSR): add a finite state machine in NewCSR (#3597)
CSR instructions used to be executed without pipelining, so a state
machine is not needed inside the CSR module. After adding an
optimization that allows certain CSRR instructions to be pipelined, a
state machine is required, since the arbiter to integer register files
must allow a write request before a CSRR instructions is successfully
executed.