fix(PageTableCache): fix hit & bypass logic when onlypf
The meaning of `onlyPf` signal is that only pagefault occurs and no other exceptions occurs. In our design intent, this case for entry l0 (4KB page) would be considered hit (by setting `vs` = 1) and the pf signal would be returned to the L1 TLB; however, the previous condition of `vs` being 1 requires that `pte.isLeaf()` must be satisfied, which requires that the `pte.v` bit is 1. In fact, for the `onlyPf` condition to be satisfied, `vs` can be set to 1 without judging `pte.isLeaf()`.

For the 8 consecutive page tables refilled into PageCache, when `pte.canRefill` and `onlyPf` are both not satisfied, `vs` is set to 0, and thus judged as hit=0. However, when the 8 entries with valid `vs` hits, the other `vs` 0's may be refilled (`refill.valid` is 1), in which case the `bypass` and `hit` signals could both be 1. This commit fixes the corner case.