fix(VMergeBuffer): vl of fof only allows setting smaller values (#3733)
The current condition, when there are no exceptions in the entry and the
frompipe carries an exception, will go directly to the next level of
determining whether it is a fof instruction. If it is a fof instruction
and a non-zero element triggers an exception, the value of vl is
modified directly.

The fof instruction is regarded as a unit-stride instruction. Therefore,
when a fof uop is split into two access operations and both of them
trigger an exception at the pipe, both of them will modify the vl value
of the same entry, which may result in a situation where a smaller vl
value is overwritten by a larger vl value.

Therefore, this modification makes a judgement when modifying the vl
value, and only allows to write the vl value which is smaller than the
current vl of the entry.