fix(dbltrp): fix sdt write and sdt/sie interaction logic
    * For the write alias of sstatus, write sdt 0 should also be blocked
by DTE close
    * For the write behavior of `mstatus/vsstatus.sdt`, since it is
directly written, the `sdt` field in `new_val` can be used to
determine whether to write `sdt` and affect `sie`. For the write
behavior of `sstatus.sdt`, due to changes in the write mask when
`DTEclose` is enabled, the original value of `sstatus.sdt` must be
considered for its impact on `sie` when `new_val.sdt` cannot be written.