fix(PTW): clear reg after resp to prevent problem
PTW use reg full_gvpn_reg to generate GPF for out-of-range gpaddr, but
this reg is not cleared after PTW response. If the last PTW request got
a bad full_gvpn_reg, the gvpn_gpf and guestFault keeps until next PTW
request, resulting in the total mess of state machine.

This patch clears full_gvpn_reg at the response of PTW. As PTW could
response in two state paths, this patch adds an independent condition to
achieve this.