fix(LoadUnit): use `lqIdx` to determine age (#4136)
1. `lqIdx` has less bit width.
2. for vectors, the `robIdx` is the same for multiple `flow`s.
Previously, for vectors, we would additionally use `uopIdx` for
judgement. But actually, in theory, we only need to use `lqIdx/sqIdx`.

Here we change the age judgement for vectors to `lqIdx` to ensure
accurate age judgement. And change the age judgement of scalar to
`lqIdx` as well to reduce the cost.