ppa(LRQ): optimise LRQ entries for ppa
Previously, due to LRQ can't backpressure, we set LoadQueueReplaySize
no smaller than VirtualLoadQueueSize to ensure uop can always enter LRQ.
But LRQ is very redundant, so it cause an increase in area and power.

Now, in order to optimize power and area, we need to reduce LRQ
redundant entries.

We should resolve this constraint at first.
Use "lqThreshold" to control the execution of load instructions.
When the free entries in LRQ reaches the "lqThreshold", we backpressure
to make IssueQueue send the oldest uop to LDU.

And then, reduce LoadQueueReplaySize from 72 to 48.