fix(L2TLB): Fix stuck caused by MissQueue full
In the previous design, we will ensure that L2 TLB will be able to
receive the request by repeater between L1 TLB and L2 TLB.
Specifically, the number of entries in L2 TLB MissQueue will be
greater than or equal to the sum of itlbRepeater entries + dtlbRepeater
entries, so as long as the repeater is able to accept, the L2 TLB will
be able to handle.

However, there is always a problem with this approach: the repeater
does not know whether multiple reqs of the TLB will be on the same huge
page (e.g. a 2MB page), so it is possible that there may be multiple
items on the same huge page in a repeater at the same time. In this
case, L2 TLB will release more than one repeater entry at a time
(these pages are in a same huge page). However, the req and resp of L2
TLB are one-to-one correspondence, although one resp has returned, the
resp of other reqs are still inside L2 TLB, which means that the
outstanding requests in L2 TLB are more than valid entries in repeater.
In this case, there is no guarantee that the L2 TLB will be able to
accept as long as the repeater is able to accept.

This problem becomes more pronounced when TLB compression (sector TLB)
is added and the repeater is reconfigured, which may result in more
situations that L2 TLB releasing more than one repeater entries at one
time. After refactoring the repeater, the three dtlb repeaters do not
check for duplicates, so L2 TLB resp may release entries from multiple
repeaters at the same time. Due to these reasons, in the high pressure
access situation especially after the hypervisor extension is
implemented, the MissQueue of L2 TLB may be filled while the repeater
cannot intercept (the repeater still has empty entries), which finally
resulting in a stuck.

This commit fixes this issue. The fix is straightforward, simply add a
counter at top level of L2 TLB to ensure that the number of outstanding
requests inside L2 TLB will not exceed the number of entries in
MissQueue. In future MMU refactoring, we should assign idx to L2 TLB
requests like freelist, so that we can solve the problem of stuckness,
and at the same time let MissQueue know the order of the requests, so
as to avoid the performance bug that the new requests are sent out
first, and the old ones are sent out later.