fix(PTW): fix gpf_level generate logic when Sv39
For historical reasons, `gpf_level` needs to be 1 greater than `af_level`. In the original design, only the Sv48 case was considered, and a similar treatment is needed for the Sv39 case.

As an extra note here, `gpf_level` is a 2-bit register, so when `gpf_level` = 0.U, `gpf_level` - 2.U = `00` - 2 = `10` = 2.U. The PTW module's `gpf_level` computation uses this feature, which is extremely unscalable, and, except for the original author, is completely incomprehensible, and is in dire need of refactoring.